Unify service contracts into GrillBot.Contracts
Every service boundary carried two hand-maintained copies of the same DTOs:
a "client" copy under GrillBot.Core.Services/<Context>/Models and a "server"
copy under the owning service's Models folder. Seven service csproj files
suppressed CS0436 so the compiler would stop complaining that it was silently
preferring the local copy.

Move both copies into a single src/Contracts/GrillBot.Contracts project, one
folder per bounded context, namespaced GrillBot.Contracts.<Context>.*. That
removes the namespace collision the suppressions were papering over, so all
seven <NoWarn>CS0436</NoWarn> entries are deleted. GrillBot.Core.Services is
reduced to the Refit client interfaces plus Common/.

Contracts keep their validation attributes and self-contained invariants.
Rules that need service state (options, the database) move into
ModelValidator<T> in the owning service, resolved by the new
ModelValidationFilter in GrillBot.Services.Common; it runs at order -3000, so
it lands ahead of [ApiController]'s model-state filter and produces the same
ValidationProblemDetails 400 as before. Domain helpers that had accreted on
the DTOs (SearchRequest.IsAdvancedFilterSet, CreateTransactionPayload
.GetIncrementType, the ImageMagick-dependent PeepoRequest members) become
extension methods in the service that owns them.

Reconciling the copies surfaced live defects:

- ApiStatistics never deserialized. The client declared ByDateInternalApi/
  ByDatePublicApi as Dictionary<string,long>; the producing service sent
  DailyInternalApi/DailyPublicApi as Dictionary<DateOnly,int>. Different
  names, so the bot read empty dictionaries. Took the producer's shape.
- LogType.MemberWarning = 26 existed only on the client; AuditLogService's
  copy stopped at 25.
- MeasuresListParams validation was split - the client copy enforced the date
  interval, the server copy enforced the Type allow-list, and neither side ran
  both. Merged.
- TransferPointsRequest had lost every attribute and the same-account rule on
  the client side. Restored.
- MessageDeletedSearchRequest.AuthorId had lost [DiscordId]/[StringLength].
- RecalculationMessage (client) published to queue "Recalculation" while its
  server twin consumed "UserRecalculation". Zero usages; deleted as dead code.
- Remind ids were long on the client and int in the entity. Unified on int.
  Discord hands slash-command integers over as long, so FinishRemind now
  range-checks into the existing not-found path rather than truncating.
- StatisticItem.SuccessRate/AvgDuration are computed rather than settable; no
  producer ever set them, so they were always zero on the wire.

AvatarInfo deliberately keeps IValidatableObject: it is only ever validated as
a nested property, and the filter resolves validators for top-level action
arguments only.

docker/deployables.json gains a "contracts" shared path wired into every .NET
deployable, so a contracts-only change rebuilds the images that depend on it.

+++ src/Bot/GrillBot.App/Actions/Commands/Reminder/CopyRemind.cs
@@ -3,7 +3,7 @@ using GrillBot.Core.Exceptions;
 using GrillBot.Core.Services.Common.Exceptions;
 using GrillBot.Core.Services.Common.Executor;
 using RemindService;
-using RemindService.Models.Request;
+using GrillBot.Contracts.Remind.Requests;
 
 namespace GrillBot.App.Actions.Commands.Reminder;
 
@@ -12,7 +12,7 @@ public class CopyRemind(
     IServiceClientExecutor<IRemindServiceClient> _remindService
 ) : CommandAction
 {
-    public async Task ProcessAsync(long originalRemindId)
+    public async Task ProcessAsync(int originalRemindId)
     {
         var request = new CopyReminderRequest
         {
