+++ src/Bot/GrillBot.App/Helpers/RemindHelper.cs
@@ -4,6 +4,6 @@ namespace GrillBot.App.Helpers;
 
 public static class RemindHelper
 {
-    public static MessageComponent CreateCopyButton(long remindId)
+    public static MessageComponent CreateCopyButton(int remindId)
         => new ComponentBuilder().WithButton(customId: $"remind_copy:{remindId}", emote: Emojis.PersonRisingHand).Build();
 }
